requirements.txt
fastapi
uvicorn
paho-mqtt
kafka-python